requirements.txt
torch>=2.0.0
transformers>=4.36.0
accelerate>=0.25.0
bitsandbytes>=0.41.0
datasets>=2.14.0
pandas>=2.0.0
numpy<2.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
shap>=0.42.0
umap-learn>=0.5.0
rouge-score>=0.1.2
tqdm>=4.65.0
pyyaml>=6.0
einops>=0.6.0
wandb>=0.15.0
